\name{hw3cca-package}
\alias{hw3cca-package}
\alias{hw3cca}
\docType{package}
\title{
\packageTitle{hw3cca}
}
\description{
\packageDescription{hw3cca}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hw3cca}
\packageIndices{hw3cca}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hw3cca}

Maintainer: \packageMaintainer{hw3cca}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
